Fix tests, refactoring
